+++ index.js
@@ -7,18 +7,17 @@ module.exports = {
     './rules/strict',
     './rules/style',
     './rules/variables',
-    './rules/promise'
+    './plugins/promise',
+    './plugins/node',
+    './plugins/import'
   ].map(require.resolve),
-
   root: true,
   noInlineConfig: true,
   reportUnusedDisableDirectives: true,
-
   globals: {
     Atomics: 'readonly',
     SharedArrayBuffer: 'readonly'
   },
-
   parserOptions: {
     ecmaVersion: 2020,
     sourceType: 'module',
@@ -26,6 +25,5 @@ module.exports = {
       impliedStrict: true
     }
   },
-
   rules: {}
 };